requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
redis
pinecone
python-multipart
pypdf
aiofiles
tavily-python
# LangChain ecosystem
langchain
langchain-core
langchain-community
langchain-google-genai
langchain_openai
langchain_pinecone
langchain_text_splitters
langchain-groq
langchain-tavily
# Vector DB (choose one)
qdrant-client
# pinecone-client

# Utilities
pydantic
